# Conflicts:
#	kurs.kz.xcodeproj/project.pbxproj

+++ kurs.kz/ExchangerDetail/Model/Details.swift
@@ -14,7 +14,7 @@ struct Details: Codable {
     let contacts: [String?]?
     let whatsapp: [String?]?
     let webSite, email: String?
-    let score, scoreCount: Int?
+    let score, scoreCount: Float?
     let open: Bool?
     
     enum CodingKeys: String, CodingKey {
